Dropped Plupload and went with AjaxControlToolkit
So it seems the choice of upload control was based on some hasty testing
and it turned out my implementation of Plupload did not fix the large
file issue (files were transferred but server-side messed up, and no
files were saved).

I tried troubleshooting Plupload, but in the process stumbled upon the
updated AjaxControlToolkit. Supplied as a NuGet package with excellent
setup guide made it hard to resist.

This also forced a update from .NET 3.5 to .NET 4.5 (merely a symbolic
update)

As per this commit everything should be functional, but deployment
testing is not finished (+4gb file takes some time to upload)